Fix build errors: exclude archive, fix types, fix test enum
- Exclude lib/_archive/** from analysis (legacy WebView code)
- Fix int→double type mismatch in DebugStatus.gridDensity
- Fix GlowIntensity.intense → GlowIntensity.engaged in tests

+++ lib/vib3/widget/vib3_widget.dart
@@ -96,7 +96,7 @@ class _VIB3WidgetState extends State<VIB3Widget>
       visualSystem: widget.system.name.toUpperCase(),
       geometryIndex: widget.geometryIndex,
       morphParameter: _state.morphParameter,
-      gridDensity: _state.tessellationDensity,
+      gridDensity: _state.tessellationDensity.toDouble(),
     );
   }
